Upload releases to s3
This borrows steps from existing providers to grab AWS credentials in order to upload the archive to `get.pulumi.com` during release.

I downgraded the goreleaser config (and action) to match what we currently use in other providers. I also added a blobs stanza, which should take care of actually publishing to S3.

This provider doesn't have a prerelease action so my plan is to merge this an cut a 1.0.1 tag to verify.

Refs https://github.com/pulumi/pulumi-terraform-provider/issues/66.